Allow Client to be Changed (#2279)
* First stab at making the client configurable

* Wrapped HTTPClient in our own wrapped type and get the tests working

* Remove commented out code

* Update Sources/Vapor/Client/Application+Client.swift


* Rename WrappedClient into AsyncHTTPClient

* Backfill test to ensure configuration can't be changed

* Make it clear in the code that config can't be changed

* Migrate client tests into their own file

* Add app.client back and expose AsyncHTTPClient's driver

* Move configuration down a layer to AsyncHTTPClient
