+++ Tests/VaporTests/ApplicationTests.swift
@@ -1016,24 +1016,6 @@ final class ApplicationTests: XCTestCase {
         }
     }
 
-    func testClientBeforeSend() throws {
-        let app = Application()
-        defer { app.shutdown() }
-        try app.boot()
-        
-        let res = try app.client.post("http://httpbin.org/anything") { req in
-            try req.content.encode(["hello": "world"])
-        }.wait()
-
-        struct HTTPBinAnything: Codable {
-            var headers: [String: String]
-            var json: [String: String]
-        }
-        let data = try res.content.decode(HTTPBinAnything.self)
-        XCTAssertEqual(data.json, ["hello": "world"])
-        XCTAssertEqual(data.headers["Content-Type"], "application/json; charset=utf-8")
-    }
-
 //    func testSingletonServiceShutdown() throws {
 //        final class Foo {
 //            var didShutdown = false
@@ -1097,34 +1079,6 @@ final class ApplicationTests: XCTestCase {
 
         let res = try app.client.get("http://127.0.0.1:8123/foo").wait()
         XCTAssertEqual(res.body?.string, "bar")
-    }
-
-    func testBoilerplateClient() throws {
-        let app = Application(.init(
-            name: "xctest",
-            arguments: ["vapor", "serve", "-b", "localhost:8080", "--log", "trace"]
-        ))
-        try LoggingSystem.bootstrap(from: &app.environment)
-        defer { app.shutdown() }
-
-        app.get("foo") { req -> EventLoopFuture<String> in
-            return req.client.get("https://httpbin.org/status/201").map { res in
-                XCTAssertEqual(res.status.code, 201)
-                req.application.running?.stop()
-                return "bar"
-            }.flatMapErrorThrowing {
-                req.application.running?.stop()
-                throw $0
-            }
-        }
-
-        try app.boot()
-        try app.start()
-
-        let res = try app.client.get("http://localhost:8080/foo").wait()
-        XCTAssertEqual(res.body?.string, "bar")
-
-        try app.running?.onStop.wait()
     }
 
     func testBoilerplate() throws {
@@ -1192,34 +1146,6 @@ final class ApplicationTests: XCTestCase {
         }
     }
 
-    func testApplicationClientThreadSafety() throws {
-        let app = Application(.testing)
-        defer { app.shutdown() }
-
-        let startingPistol = DispatchGroup()
-        startingPistol.enter()
-        startingPistol.enter()
-
-        let finishLine = DispatchGroup()
-        finishLine.enter()
-        Thread.async {
-            startingPistol.leave()
-            startingPistol.wait()
-            XCTAssert(type(of: app.client.http) == HTTPClient.self)
-            finishLine.leave()
-        }
-
-        finishLine.enter()
-        Thread.async {
-            startingPistol.leave()
-            startingPistol.wait()
-            XCTAssert(type(of: app.client.http) == HTTPClient.self)
-            finishLine.leave()
-        }
-
-        finishLine.wait()
-    }
-
     func testRequestRemoteAddress() throws {
         let app = Application(.testing)
         defer { app.shutdown() }
@@ -1231,36 +1157,6 @@ final class ApplicationTests: XCTestCase {
         try app.testable(method: .running).test(.GET, "remote") { res in
             XCTAssertContains(res.body.string, "IP")
         }
-    }
-
-    func testClientConfigurationChange() throws {
-        let app = Application(.testing)
-        defer { app.shutdown() }
-
-        app.client.configuration.redirectConfiguration = .disallow
-
-        app.get("redirect") {
-            $0.redirect(to: "foo")
-        }
-
-        let server = try app.server.start(hostname: "localhost", port: 8080)
-        defer { server.shutdown() }
-
-        let res = try app.client.get("http://localhost:8080/redirect").wait()
-
-        XCTAssertEqual(res.status, .seeOther)
-    }
-
-    func testClientResponseCodable() throws {
-        let app = Application(.testing)
-        defer { app.shutdown() }
-
-        let res = try app.client.get("https://httpbin.org/json").wait()
-
-        let encoded = try JSONEncoder().encode(res)
-        let decoded = try JSONDecoder().decode(ClientResponse.self, from: encoded)
-        
-        XCTAssertEqual(res, decoded)
     }
     
     func testMultipleChunkBody() throws {
@@ -1298,11 +1194,13 @@ final class ApplicationTests: XCTestCase {
             defer { current += 1 }
             return Test(number: current)
         }
+        
+        app.clients.use(.responder)
 
         let cache = EndpointCache<Test>(uri: "/number")
         do {
             let test = try cache.get(
-                using: app.responder.client,
+                using: app.client,
                 logger: app.logger,
                 on: app.eventLoopGroup.next()
             ).wait()
@@ -1310,7 +1208,7 @@ final class ApplicationTests: XCTestCase {
         }
         do {
             let test = try cache.get(
-                using: app.responder.client,
+                using: app.client,
                 logger: app.logger,
                 on: app.eventLoopGroup.next()
             ).wait()
@@ -1326,6 +1224,8 @@ final class ApplicationTests: XCTestCase {
         struct Test: Content {
             let number: Int
         }
+        
+        app.clients.use(.responder)
 
         app.get("number") { req -> Response in
             defer { current += 1 }
@@ -1338,7 +1238,7 @@ final class ApplicationTests: XCTestCase {
         let cache = EndpointCache<Test>(uri: "/number")
         do {
             let test = try cache.get(
-                using: app.responder.client,
+                using: app.client,
                 logger: app.logger,
                 on: app.eventLoopGroup.next()
             ).wait()
@@ -1346,7 +1246,7 @@ final class ApplicationTests: XCTestCase {
         }
         do {
             let test = try cache.get(
-                using: app.responder.client,
+                using: app.client,
                 logger: app.logger,
                 on: app.eventLoopGroup.next()
             ).wait()
@@ -1356,7 +1256,7 @@ final class ApplicationTests: XCTestCase {
         sleep(1)
         do {
             let test = try cache.get(
-                using: app.responder.client,
+                using: app.client,
                 logger: app.logger,
                 on: app.eventLoopGroup.next()
             ).wait()
@@ -1491,19 +1391,12 @@ final class ApplicationTests: XCTestCase {
     }
 }
 
-extension Application.Responder {
-    /// Creates a `Client` from an `Application`'s  current`Responder`.
-    var client: Client {
-        ResponderClient(responder: self.current, application: self.application)
-    }
-}
-
 struct ResponderClient: Client {
     let responder: Responder
     let application: Application
 
-    var eventLoopGroup: EventLoopGroup {
-        self.application.eventLoopGroup
+    var eventLoop: EventLoop {
+        self.application.eventLoopGroup.next()
     }
 
     func `for`(_ request: Request) -> Client {
@@ -1526,8 +1419,20 @@ struct ResponderClient: Client {
             ClientResponse(status: res.status, headers: res.headers, body: res.body.buffer)
         }
     }
+}
 
+extension Application.Clients.Provider {
+    static var responder: Self {
+        .init {
+            $0.clients.use { $0.clients.responder }
+        }
+    }
+}
 
+extension Application.Clients {
+    var responder: ResponderClient {
+        return ResponderClient(responder: self.application.responder, application: self.application)
+    }
 }
 
 private extension ByteBuffer {
